openai-chatkit-fasthtml/requirements.txt
python-fasthtml>=0.12
python-dotenv>=1.0.0
httpx>=0.25.0
uvicorn>=0.24.0
pydantic>=2.0.0
